refactor: created app module
