+++ src/osparc_function_services/cli.py
@@ -1,6 +1,6 @@
 import typer
 
-from ._utils import run_as_service
+from .app import run_as_service
 from .sensitivity_ua_services import (
     sensitivity_ua_linear_regression,
     sensitivity_ua_test_func,
